FEAT: add ci
This adds ci for pull requests however this drops the tsc --noEmit in
build. This is unfortunately due to the fact upstream's types aren't as
strict, and use playright extensions to window. Instead of fighting to
try and fix tsconfig to support them(going against what upstream does)
we drop usage of tsc --noEmit